// lib/types/futures.ts
export type PositionSide = 'BOTH' | 'LONG' | 'SHORT';

export type MarginType = 'ISOLATED' | 'CROSSED';

export type WorkingType = 'MARK_PRICE' | 'CONTRACT_PRICE';

export type FuturesOrderType =
  | 'LIMIT'
  | 'MARKET'
  | 'STOP'
  | 'STOP_MARKET'
  | 'TAKE_PROFIT'
  | 'TAKE_PROFIT_MARKET'
  | 'TRAILING_STOP_MARKET';

// When using the submitMultipleOrders() endpoint, it seems to expect strings instead of numbers. All other endpoints use numbers.
export interface NewFuturesOrderParams<numberType = number> {
  [x: string]: string;
  symbol: string;
  side: OrderSide;
  positionSide?: PositionSide;
  type: FuturesOrderType;
  timeInForce?: OrderTimeInForce;
  quantity?: numberType & string;
  reduceOnly?: BooleanString;
  price?: numberType & string;
  newClientOrderId?: string;
  stopPrice?: numberType & string;
  closePosition?: BooleanString;
  activationPrice?: numberType & string;
  callbackRate?: numberType & string;
  workingType?: WorkingType;
  priceProtect?: BooleanStringCapitalised;
  newOrderRespType?: OrderResponseType;
}
// ...
export interface FuturesPosition {
  entryPrice: numberInString;
  marginType: Lowercase<MarginType>;
  isAutoAddMargin: 'false' | 'true';
  isolatedMargin: numberInString;
  leverage: numberInString;
  liquidationPrice: numberInString;
  markPrice: numberInString;
  maxNotionalValue: numberInString;
  positionAmt: numberInString;
  notional: numberInString;
  isolatedWallet: numberInString;
  symbol: string;
  unRealizedProfit: numberInString;
  positionSide: PositionSide;
  updateTime: number;
}